Foundation
Made a working run of the lyric cloud front end using generic html and
css in one case, and using foundation in the other. The foundation
implement is implemented, the generic html and css is commented outside
the html tag. Facebook share button needs to be implemented, a
placeholder button is there. The lyriccloud generation needs to be
implemented, but the div space is there for when its made. The input
textfield is ready and is in a form for submission to query the API and
stuff.  Made changes in the CSS of foundation to get buttons to default
to purple not blue. Made some default CSS classes and things.